bandwidth plugin: split unit tests in several files
